A hack to make pyunifiprotect work with UniFi OS 3.1.16.
This is related to AngellusMortis/pyunifiprotect#307. It looks like some events
still have the original `modelKey` while some have `modelkey`, all lower case.

The `modelKey` is referred to in so many places that making all of them check
for `modelkey` additionally would be a rather ugly fix.

One solution could've been be to convert the JSON objects into CaseInsensitiveDicts
instead of dicts, but the orjson library doesn't support such mechanisms.